# openapi.yml
openapi: 3.0.3
info:
  title: ACMS - Apartment Committee Management System
  description: |
    Multi-tenant SaaS platform for apartment/housing societies.
    Five role types: site_admin, apartment_admin, committee_head, committee_member, resident.
    Tenant isolation via `apartmentId` — all non-site-admin queries are scoped.
  version: 1.0.0
  contact:
    name: ACMS Team
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApartmentStatus:
      type: string
      enum: [active, suspended]
    PlanPriceType:
      type: string
      enum: [per-unit, flat]
...
    InvoiceStatus:
      type: string
      enum: [paid, unpaid]
...
    Apartment:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        address: { type: string }
        status: { $ref: '#/components/schemas/ApartmentStatus' }
        planId: { type: string, nullable: true, description: 'Plan ID or populated object' }
        createdAt: { type: string, format: date-time }
    CreateApartmentRequest:
      type: object
      required: [name, address]
      properties:
        name: { type: string, description: Unique apartment name (lowercased) }
        address: { type: string }
    UpdateApartmentRequest:
      type: object
      properties:
        address: { type: string }
        status: { $ref: '#/components/schemas/ApartmentStatus' }
        planId: { type: string, nullable: true }
    CreateApartmentAdminRequest:
      type: object
      required: [apartmentId, name, identifier, password]
      properties:
        apartmentId: { type: string }
        name: { type: string }
        identifier: { type: string, description: Login username }
        password: { type: string, minLength: 6 }
      example:
        apartmentId: 664abc...
        name: John Doe
        identifier: johndoe
        password: securepass123

    # ────────── Plan ──────────
    Plan:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        priceType: { $ref: '#/components/schemas/PlanPriceType' }
        price: { type: number }
        billingCycle: { type: string, default: monthly }
        features: { type: array, items: { type: string } }
    CreatePlanRequest:
      type: object
      required: [name, priceType, price]
      properties:
        name: { type: string }
        priceType: { $ref: '#/components/schemas/PlanPriceType' }
        price: { type: number, exclusiveMinimum: 0 }
        features: { type: array, items: { type: string } }

    # ────────── SaaS Invoice ──────────
    SaaSInvoice:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string, description: Apartment ID or populated object }
        planId: { type: string, description: Plan ID or populated object }
        period: { type: string, example: '2026-07' }
        amount: { type: number }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        generatedAt: { type: string, format: date-time }
        dueDate: { type: string, format: date-time }
    MarkInvoicePaidRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [paid] }

    # ────────── Unit ──────────
...
    Notice:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        committeeId: { type: string, nullable: true, description: 'Null = apartment-wide' }
        title: { type: string }
        body: { type: string }
        postedBy: { type: string, description: 'User ID or populated' }
        createdAt: { type: string, format: date-time }
    CreateNoticeRequest:
      type: object
      required: [title, body]
      properties:
        committeeId: { type: string, nullable: true, description: 'Omit for apartment-wide' }
        title: { type: string }
        body: { type: string }

    # ────────── Error ──────────
...
paths:
...
  /admin/apartments:
    get:
      tags: [Site Admin]
      summary: List all apartments (with plan details)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of apartments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Apartment'
    post:
      tags: [Site Admin]
      summary: Create a new apartment
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApartmentRequest'
      responses:
        '201': { description: Apartment created }
        '409': { description: Apartment name already exists }
  /admin/apartments/{id}:
    put:
      tags: [Site Admin]
      summary: Update apartment (address, status, plan assignment)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApartmentRequest'
      responses:
        '200': { description: Apartment updated }
        '404': { description: Apartment not found }
    delete:
      tags: [Site Admin]
      summary: Delete an apartment
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Apartment deleted }
  /admin/apartment-admins:
    post:
      tags: [Site Admin]
      summary: Create an Apartment Admin account for a given apartment
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApartmentAdminRequest'
      responses:
        '201': { description: Apartment admin created }
        '400': { description: Type must be apartment_admin }
        '404': { description: Apartment not found }
        '409': { description: Identifier already exists in this apartment }

  # ══════════════════════════════════════════════════
  # SITE ADMIN — Plans
  # ══════════════════════════════════════════════════
  /admin/plans:
    get:
      tags: [Site Admin]
      summary: List all pricing plans
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'
    post:
      tags: [Site Admin]
      summary: Create a pricing plan
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201': { description: Plan created }
  /admin/plans/{id}:
    put:
      tags: [Site Admin]
      summary: Update a plan
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '200': { description: Plan updated }
        '404': { description: Plan not found }
    delete:
      tags: [Site Admin]
      summary: Delete a plan
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Plan deleted }

  # ══════════════════════════════════════════════════
  # SITE ADMIN — SaaS Invoices
  # ══════════════════════════════════════════════════
  /admin/invoices:
    get:
      tags: [Site Admin]
      summary: List all SaaS invoices across all apartments
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaaSInvoice'
  /admin/invoices/generate:
    get:
      tags: [Site Admin]
      summary: Manually trigger monthly invoice generation
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Generation results
  /admin/invoices/{id}/mark-paid:
    put:
      tags: [Site Admin]
      summary: Mark an invoice as paid
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkInvoicePaidRequest'
      responses:
        '200': { description: Invoice updated }
        '404': { description: Invoice not found }

  # ══════════════════════════════════════════════════
  # APARTMENT ADMIN — Invoices (own apartment)
  # ══════════════════════════════════════════════════
...
  /notices:
    get:
      tags: [Notices]
      summary: List notices (scoped by role — committee sees relevant + apartment-wide)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notice'
    post:
      tags: [Notices]
      summary: Post a notice (apartment_admin or committee_head)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoticeRequest'
      responses:
        '201': { description: Notice created }
        '403': { description: Residents/members cannot post notices }
  /notices/{id}:
    delete:
      tags: [Notices]
      summary: Delete own notice
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Notice deleted }
        '404': { description: Not found or unauthorized }

  # ══════════════════════════════════════════════════
  # DASHBOARDS
  # ══════════════════════════════════════════════════
  /dashboard/site-admin:
    get:
      tags: [Dashboards]
      summary: Site admin overview (counts, revenue, recent records)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Dashboard data
  /dashboard/apartment-admin:
    get:
      tags: [Dashboards]
      summary: Apartment admin overview (units, residents, committees, finances)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Dashboard data
  /dashboard/committee-head:
    get:
      tags: [Dashboards]
      summary: Committee head (and member) overview
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Dashboard data
  /dashboard/resident:
    get:
      tags: [Dashboards]
      summary: Resident dashboard (bills, complaints, notices for their unit)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Dashboard data

  # ══════════════════════════════════════════════════
  # HEALTH
  # ══════════════════════════════════════════════════
...